feat: set up build system
Finally figured out how to get both the core building of firmware with
the building and execution of unit tests on my native system. Took a
bit, but figured out that using a separate makefile in the `tests`
directory is one solution that avoids conflicts with the gcc and avr-gcc
toolchains when construction object and ELF files.

I made sure to make the Makefiles as portable as I possible could get
them. I ran them through both GNU Make and pdpmake as a way of testing
their cross-platform capabilities. I have yet to see if these makefiles
work across FreeBSD.

The code that was compiled is rather simple:

- The code in `src/main.c` just interacts with USART through polling.
- The code in `src/ringbuf.c` implements a very basic ring buffer
  container.
- The code in `tests/*.c` simple sets up a basic unit test to see if
  the current build system can actually build with the `acutest`
  library.

Overall, this is a good first step towards the development of the RTC
scheduling firmware.

Signed-off-by: Jason Pena <jasonpena@awkless.com>