refactor(ringbuf): make API ISR compatible
Now the ring buffer API can be used with interrupts for the Atmega328p.
I also removed the `size` member, because I can just use the `head` and
`tail` members to determine the bounds of the buffer with
`RINGBUF_CAPACITY`.

I also made it easier to determine with the ring buffer overflows
or underflows with negative error codes.

Signed-off-by: Jason Pena <jasonpena@awkless.com>